EchoMood-Backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
motor          # async MongoDB driver
python-dotenv  # load .env
